backend/requirements.txt
# requirements.txt

# Основные зависимости FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
loguru==0.7.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9

# Для работы с Excel и данными
pandas>=1.5.0
openpyxl>=3.0.0

# Для парсера врачей
pyppeteer>=1.0.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
lxml>=4.9.0

# Дополнительные зависимости для парсера
requests>=2.28.0
